Add warning for &default params not in declarations
Using &default on a function/hook/event parameter has no effect, so
there's now a warning for that.  Otherwise, there's not currently a way
to detect ambiguities should one use different &default expressions
between definitions, and they may be surprised or else silently get
wrong/unintended behavior and not notice in that case.