&expire_func(table, index1, index2, ...)
Allows expire func to take multiple indexes corresponding to the
number of indexes in the table that it is associated with. Also adds
type checking for those arguments.

Previously there existed a work around to this where expire_func
could take a value of TYPE_ANY as an argument and then do tuple
unpacking to get the index values. This was used with some frequency
inside zeek so this retains that backwards functionality.