Add first-class support for overloaded functions
Resolving an overload function into a Val now stores both the Func
object along with an overload index.  Generally, when we assign an
overloaded function name to some variable, return it from a function,
or pass it into another function, we can resolve it into a particular
overload at parse-time since we know the target type.

There's still a couple cases where we might not be able to resolve
a function overload at parse-time:

* When assigning an overloaded name into an `any`, we wait until
  performing a call to find the correct overload, if any

* When referencing an overloaded function, we'll generally know
  the overload type/index, but maybe not the particular Func object
  since the function may be defined later or even assigned at run-time.
  So there's a new expression type for that: FuncRefExpr