Stop incorrect language identification on github.
If you look at the languages that github is implemented in according
to our page on github, it has several incorrect things. PAC files and
SIG files being the big misidentifications. In my opinion it would look
better to mark these as no language so that javascript and ML stop showing
up as languages that Zeek is implemented in.

This change should make fix that on github according to:
	https://github.com/github/linguist#overrides