Change reassembly data structure to use a skip-list
Switching from a plain doubly linked-list to a skip-list improves the
O(n) overlap detection in the case where one endpoint is no longer
ACKing data, but we've been tracking the sent-data for a while and
there happens to be duplicate segments near the end of the list.

The worse case overlap detection complexity for a skip-list will
still be O(n), but its probabilistic properties put the expected
average at O(log n).