Remove Zeek-side batching of Broker messages
The Zeek-side batching was implemented in response to observing
performance degradation with Broker compared to the old communication
backend. This seems no longer necessary, so we can remove the opaque
data as well as the batching messages in Broker (see
https://github.com/zeek/broker/pull/64).

Instead of serializing into opaque buffers, we simply convert the
threading values directly to `broker::data` and vice versa.