main: drop privileges after initialization
Zeek is designed to run as root (to be able to capture packets).  This
makes Zeek itself a huge surface for attackers.  Zeek parses and
evaluates malicious data, after all.

As a proactive countermeasure, this patch drops unneeded privileges to
reduce the damage from a hack:

- create a new mount namespace, unmount various filesystems and
  remount others read-only
- mount a private tmpfs on /tmp, /var/tmp and /dev/shm
- set the "NO_NEW_PRIVS" process flag

All of this is Linux specific.  Other operating systems have other
knobs which may be interesting for Zeek.

Though much more needs to be done to make Zeek more secure, for
example:

- create a user namespace (see code comment)
- change to a different uid/gid
- install a seccomp filter to create a whitelist of allowed system
  calls
- run Zeek in an empty network namespace so attackers cannot
  communicate with the internet
- make more parts of the filesystem unavailable, e.g. /var, /opt, /etc
  or better: run Zeek in an empty VFS where just the output directory
  is bind-mounted
- run the dangerous packet evaluation code in a dedicated process
  which has a socket to the main process and nothing else, no
  filesystem, no network, a tiny system call whitelist (Privilege
  Separation)

You can't be paranoid enough with software like Zeek!