Store updates: downlevel and store table elements as keys.
In this commit we change our approach and just store table keys as
broker keys and table values as broker indexes. This means we only can
have a single table in a broker store.

This seems to work perfectly - and leads to vastly less complex code.

Code should work, but is actually pretty untested. Might break with
certain types. Does not yet handle item expiration.

Also - this has the tiny issue that all remote operations currently lead
to loops (the operation will immediately be sent to the broker-store
again), which is not quite optimal.